Backend/requirements.txt
# Django Core
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Async & Performance
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
channels>=4.0.0
channels-redis>=4.1.0

# Database
psycopg2-binary>=2.9.9
django-environ>=0.11.0

# Google Cloud Storage
google-cloud-storage>=2.10.0
django-storages>=1.14.0

# GeoIP for region detection
geoip2>=4.7.0

# Email
django-ses>=3.5.0
aiosmtplib>=3.0.0

# Security & Rate Limiting
django-ratelimit>=4.1.0
djangorestframework-simplejwt>=5.3.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3

# Development
django-debug-toolbar>=4.2.0

